# Conflicts:
#	dds/DCPS/DiscoveryBase.h
#	dds/DCPS/MessageTracker.cpp
#	dds/DCPS/ReactorTask.cpp
#	dds/DCPS/transport/framework/QueueTaskBase_T.h

+++ tests/DCPS/Messenger/Args.h
@@ -21,7 +21,7 @@
 #include <iostream>
 #include <cstdlib>
 
-const int num_messages = 40;
+const size_t num_messages = 40;
 extern bool reliable;
 
 inline int
